tracking-manager/packages/tracking-api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic~=2.10.5
python-dotenv~=1.0.1
pika==1.3.2
redis~=5.2.1
motor==3.3.2
watchdog==3.0.0
python-multipart==0.0.6
aiofiles==23.2.1
boto3~=1.35.97
python-magic==0.4.27

typing~=3.7.4.3
pymongo~=4.9.2
httpx~=0.28.1
bcrypt~=4.2.1
PyJWT~=2.10.1
elasticsearch~=7.10
pandas~=2.2.3
pyfa_converter_v2==2.0.0rc1
botocore~=1.35.97
opensearch-py~=2.8.0
resend~=2.6.0
requests~=2.32.3